Rename py-api folder to py_api
