+++ scripts/docker/start-py.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo "Starting Python API..."
-exec python3 src/py-api/app.py
+exec python3 src/py_api/app.py
