+++ tests/py-api/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 from fastapi.testclient import TestClient
-from src.py-api.app import app
+from src.py_api.app import app
 
 @pytest.fixture
 def client():
