+++ tests/py-api/unit/test_helper_module.py
@@ -1,6 +1,5 @@
-import pytest
 import datetime
-from src.py-api.helper_module import *
+from src.py_api.helper_module import *
 
 # TODO: Add tests for type checking and error handling
